user specific page

+++ client/src/components/HomePage/index.jsx
@@ -5,14 +5,27 @@ const apiUrl = import.meta.env.VITE_API_URL;
 const TaskManager = () => {
   const [tasks, setTasks] = useState([]);
   const [taskInput, setTaskInput] = useState('');
+  const [userName, setUserName] = useState('Task Manager')
 
   useEffect(() => {
     fetchTasks();
+    fetchUserName();
   }, []);
 
   const fetchTasks = async () => {
     try {
-      const response = await fetch(`${apiUrl}/tasks`);
+      const token = localStorage.getItem('token'); // Retrieve the authentication token
+      if (!token) {
+        return response
+          .status(401)
+          .json({ success: false, msg: "Token missing!" });
+      }
+
+      const response = await fetch(`${apiUrl}/tasks`, {
+        headers: {
+          'Authorization': `Bearer ${token}`, // Include the token in the request headers
+        },
+      });
       const data = await response.json();
       setTasks(data.tasks);
     } catch (error) {
@@ -20,12 +33,39 @@ const TaskManager = () => {
     }
   };
 
+  const fetchUserName = async () => {
+    try {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        throw new Error("Token missing!"); 
+      }
+
+      const response = await fetch(`${apiUrl}/users/me`, {
+        headers: {
+          'Authorization': `Bearer ${token}`,
+        },
+      });
+      console.log(response);
+      const data = await response.json();
+      console.log(data);
+      setUserName(`${data.firstName}`);
+    } catch (error) {
+      console.error('Error fetching user name:', error);
+    }
+  };
+
   const createTask = async () => {
     try {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        throw new Error("Token missing!"); 
+      }
+
       const response = await fetch(`${apiUrl}/tasks`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`, // Include the token in the request headers
         },
         body: JSON.stringify({ name: taskInput }),
       });
@@ -40,10 +80,15 @@ const TaskManager = () => {
 
   const updateTask = async (taskId, updatedName) => {
     try {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        throw new Error("Token missing!"); 
+      }
       const response = await fetch(`${apiUrl}/tasks/${taskId}`, {
         method: 'PATCH',
         headers: {
           'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
         },
         body: JSON.stringify({ name: updatedName }),
       });
@@ -57,8 +102,16 @@ const TaskManager = () => {
 
   const deleteTask = async (taskId) => {
     try {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        throw new Error("Token missing!"); 
+      }
+
       const response = await fetch(`${apiUrl}/tasks/${taskId}`, {
         method: 'DELETE',
+        headers: {
+          'Authorization': `Bearer ${token}`
+        },
       });
 
       await response.json();
@@ -82,7 +135,7 @@ const TaskManager = () => {
 
   return (
     <div className={styles.body}>
-      <h2 id={styles.h2}>Task Manager</h2>
+      <h2 id={styles.h2}>{userName + "'s TaskManager"}</h2>
       <form>
         <input
           type="text"
